Fix failure on server creation
Creating a server requires to have set either the location or the datacenter but not both.
This refactor changes the CreateServer constructor such that both fields are nullable but will definitely be set by the secondary constructor.
It also now includes a more precise way to set the required field by using sealed classes.